backend/requirements.txt
attrs==23.1.0
blinker==1.6.2
click==8.1.6
docker-interface==0.4.3
flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
numpy==1.26.2
PyYAML==6.0.1
referencing==0.31.1
rpds-py==0.13.2
scikit-learn==1.4.1.post1
scipy==1.11.4
sklearn==0.0
threadpoolctl==3.2.0
werkzeug==3.0.0
opencv-python==4.9.0.80
keras==2.15.0
pillow==10.3.0
tensorflow==2.15.1